requirements.txt
ultralytics
opencv-python
matplotlib
google-genai
python-dotenv
